http_chunk data handling/passing improvement.
This is a modification to the http_chunk module to forward any full
chunk received, regardless of whether the size field for the following
chunk has been received yet.  This allows http_chunk to be used in
situations where a long term HTTP connection is used to send periodic
status updates as individual chunks.  Previously a given chunk would
not be forwarded to the client process until the size for the next
chunk had been read which rendered the module difficult to use for
the scenario described.