Workaround broken ODBC implementation on OSX 10.5
There is a bug with the ODBC implementation on OSX 10.5 that manifests
itself by an ODBC application throwing a bus error when
SQLSetConnectAttr() is called on a connection that has not been previously
been opened via SQLConnect().

The change in this patch is to change when the SQLSetConnectAttr() calls
are done. Now they are only done after the SQLConnect() call succeeds.
This should not change any behaviour because, as can be seen at
http://msdn.microsoft.com/en-us/library/ms713605(VS.85).aspx ,
all the relevant attributes can be set either before or after the call
to SQLConnect() (the second table on the page).

Without this patch, any call to odbc:connect() on OSX 10.5 will fail
because the odbcserver program throws a bus error the instant it encounters
a call to SQLSetConnectAttr().