compiler: Teach 'slim' to omit compilation info
The original intention for the undocumented 'slim' option was
to omit non-essential part of the *.beam files to reduce the
size of the primary bootstrap. Therefore, debug information and
local function (only used by xref, not by loader) are omitted,
but information about the compilation time and compiler version
are still included.

Including compilation information is troublesome, however, when
commit the boostrap into a revision control system, because every
beam file is guaranteed to be changed every time the bootstrap is
updated.

Therefore, change the 'slim' optioin to also omit compilation
information.