Fix inets HTTP client bug crafting Host header when port is not 80
The host header should include the port number as well as the host
name when making a request to a server listening on a port other
than the HTTP default of 80. Currently, only the host name is
included. This is important to make the http client more
compliant with the HTTP specification. The problem can be provoked
by starting a local server listening on port 8080, starting the
erlang shell and running:
http:request(get, {"http://localhost:8080", []}, [], []).
Capture and inspection of the output from a program such as tcpdump
shows the Host header missing the port number.