change the examples' -include to -include_lib with ct.hrl
The current templates in the documentation examples include the file
ct.hrl with the module attribute -include. The user would then need
to manually pass the include path to the compiler to have it
working. This is not needed as common_test is part of the standard
OTP distribution and should work with -include_lib.