add options to binary_to_term
binary_to_term is a powerful tool for easily shuttling data over
the network.  It is not suitable for use with untrusted data, as it
happily leaks atoms and transfers data types that can carry
arbitrary data into your code (funs, pids, refs).

After a short discussion on the list, consensus appears to be that
binary_to_term needed options.  This patch implements the second
form of binary_to_term that takes options, and a simple option
called 'safe'.  If specified, this option will refuse to decode
any dangerous data types and will not leak atoms.